Refactor codes, modify name and organize files
